It was funny on HTTPS
On HTTPS the font style sheet doesn't load because it is a HTTP resource. As the result the fonts are rendered in the default cursive font, usually Comic Sans MS for some browser/OS setups. Fixed this by changing the protocol.

Also simplified it by loading just one style sheet.